fix(src): Фикс рассинхрона идентификатора на клиенте и сервере
