+++ Content.Client/Imperial/Medieval/Identity/MedievalIdentitySystem.cs
@@ -1,0 +1,7 @@
+using Content.Shared.Imperial.Medieval.IdentityManagement;
+
+namespace Content.Client.Imperial.Medieval.Identity;
+
+public sealed class MedievalIdentitySystem : SharedMedievalIdentitySystem
+{
+}
